Added new hook - "useFilterModal.ts" with all the states from FilterModal.tsx and used it in FilterModal.tsx for better code readability.
